# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CS2KZ API",
    "description": "Source Code available on [GitHub](https://github.com/KZGlobalTeam/cs2kz-api).",
    "contact": {
      "name": "AlphaKeks",
      "email": "alphakeks@dawn.sh"
    },
    "license": {
      "name": "Licensed under the GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with Steam.",
        "description": "This will redirect the user to Steam, where they can login. A session for\nthem will be created and they're redirected back to `redirect_to`.",
        "operationId": "login",
        "parameters": [
          {
            "name": "redirect_to",
            "in": "query",
            "description": "URL to redirect the user back to after a successful login.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "303": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout again.",
        "description": "This will invalidate your current session, and potentially every other\nsession as well (if `invalidate_all_sessions=true` is specified).",
        "operationId": "logout",
        "parameters": [
          {
            "name": "invalidate_all_sessions",
            "in": "query",
            "description": "Whether to invalidate all (still valid) sessions of this user.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "303": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": []
          }
        ]
      }
    },
    "/auth/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "The endpoint hit by Steam after a successful login.",
        "description": "This should not be used directly, and trying to do so will lead to errors.",
        "operationId": "callback",
        "parameters": [
          {
            "name": "openid.ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "openid.claimed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "openid.mode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.return_to",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "openid.op_endpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.response_nonce",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.invalidate_handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "openid.assoc_handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.signed",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.sig",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "default": null,
                  "nullable": true
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
...
    "/admins/{steam_id}": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "Fetch a specific admin by their SteamID.",
        "operationId": "get",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A KZ admin.",
                  "required": [
                    "name",
                    "steam_id",
                    "permissions"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "The admin's name."
                    },
                    "steam_id": {
                      "$ref": "#/components/schemas/SteamID"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "The admin's permissions.",
                      "example": [
                        "bans",
                        "servers"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Admins"
        ],
        "summary": "Create/Update an admin.",
        "description": "This endpoint is idempotent!",
        "operationId": "put",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "admins"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUpdate": {
        "type": "object",
        "description": "Request payload for updating an admin.",
        "required": [
          "permissions"
        ],
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "New set of permissions for the admin.",
            "example": [
              "bans",
              "servers"
            ]
          }
        }
      },
...
      "SteamID": {
        "oneOf": [
          {
            "type": "string",
            "title": "Steam ID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Steam ID3",
            "example": "U:1:322356345"
          },
          {
            "type": "integer",
            "title": "Steam ID32",
            "example": 322356345
          },
          {
            "type": "integer",
            "title": "Steam ID64",
            "example": 76561198282622073
          }
        ],
        "description": "a player's SteamID",
        "example": "STEAM_1:1:161178172"
      },
...
    },
    "securitySchemes": {
      "API Key": {
        "type": "http",
        "scheme": "bearer"
      },
      "Browser Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kz-auth"
      },
      "CS2 Server": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}